feat: wip farmer content tool integration
