+++ apps/picsa-apps/extension-app/src/app/routes/index.ts
@@ -1,0 +1,2 @@
+export * from './app-routes';
+export * from './tool-routes';
