+++ tsconfig.base.json
@@ -31,7 +31,7 @@
       "@picsa/data": ["libs/data/index.ts"],
       "@picsa/environments": ["libs/environments/src/index.ts"],
       "@picsa/extension-content/*": ["apps/picsa-tools/extension-content/*"],
-      "@picsa/extension/*": ["apps/native/extension-toolkit/*"],
+      "@picsa/extension/*": ["apps/picsa-apps/extension-app/*"],
       "@picsa/farmer-activity/*": ["apps/picsa-tools/farmer-activity/*"],
       "@picsa/farmer-content/*": ["apps/picsa-tools/farmer-content/*"],
       "@picsa/forms": ["libs/forms"],
